requirements.txt
qrcode[pil]
pillow
reportlab
svglib
